Convert uses of defadvice into define-advice
As of Emacs 30.1, defadvice is deprecated. Although it still works,
loading smart-tabs-mode will now cause an annoying warning in the
messages buffer.

This patch rewrites the uses of defadvice into uses of define-advice
instead. define-advice is a helper function that combines defun with
advice-add, which is the replacement of defadvice. No changes in
semantics is expected from this rewrite.